+++ src/classes/folders/TssFolderAuditSummary.class.ps1
@@ -1,0 +1,12 @@
+﻿class TssFolderAuditSummary {
+    # Action
+    [string]$Action
+    # Audit Folder Id
+    [int]$AuditFolderId
+    # Data recorded
+    [datetime]$DateRecorded
+    # Display Name
+    [string]$DisplayName
+    # Notes
+    [string]$Notes
+}