+++ src/functions/folders/Get-FolderAudit.ps1
@@ -1,0 +1,65 @@
+﻿function Get-FolderAudit {
+    <#
+    .SYNOPSIS
+    Get a list of audits
+
+    .DESCRIPTION
+    Get a list of audit for Folder ID
+
+    .EXAMPLE
+    PS> $session = New-TssSession -SecretServer https://alpha -Credential $ssCred
+    PS> Get-TssFolderAudit -TssSession $session -Id 42
+
+    Gets the audit entries for Folder ID
+
+    .NOTES
+    Requires TssSession object returned by New-TssSession
+    #>
+    [CmdletBinding()]
+    [OutputType('TssFolderAuditSummary')]
+    param (
+        # TssSession object created by New-TssSession for auth
+        [Parameter(Mandatory,
+            ValueFromPipeline,
+            Position = 0)]
+        [TssSession]$TssSession,
+
+        # Short description for parameter
+        [Parameter(Mandatory,ValueFromPipelineByPropertyName)]
+        [Alias("FolderAuditId")]
+        [int[]]
+        $Id
+    )
+    begin {
+        $tssParams = $PSBoundParameters
+        $invokeParams = @{ }
+    }
+
+    process {
+        Write-Verbose "Provided command parameters: $(. $GetInvocation $PSCmdlet.MyInvocation)"
+        if ($tssParams.ContainsKey('TssSession') -and $TssSession.IsValidSession()) {
+            foreach ($folder in $Id) {
+                $restResponse = $null
+                $uri = $TssSession.ApiUrl, 'folders', $folder, 'audit' -join '/'
+                $invokeParams.Uri = $uri
+                $invokeParams.Method = 'GET'
+
+                $invokeParams.PersonalAccessToken = $TssSession.AccessToken
+                Write-Verbose "$($invokeParams.Method) $uri"
+                try {
+                    $restResponse = Invoke-TssRestApi @invokeParams
+                } catch {
+                    Write-Warning "Issue getting folder [$folder]"
+                    $err = $_
+                    . $ErrorHandling $err
+                }
+
+                if ($restResponse) {
+                    . $TssFolderAuditSummaryObject $restResponse.records
+                }
+            }
+        } else {
+            Write-Warning "No valid session found"
+        }
+    }
+}